+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 - Fixes an issue where `emulators:exec` would return a 0 error code when emulators failed to start. (#7974)
+- Added `--import` and `emulators:export` support to the Data Connect emulator.
+- Added `firebase.json#emulators.dataconnect.dataDir`. When set, Data Connect data will be persisted to the configured directory between emulator runs.